Move Dialog to storybook(#149)
