+++ app/.storybook/preview.js
@@ -3,13 +3,16 @@ import "../src/assets/styles/globals.css";
 import { ThemeContextProvider } from "../src/components/core/theme";
 import { BrowserRouter } from "react-router-dom";
 import { NotificationProvider } from "../src/hooks/useNotification";
+import {DialogProvider} from "../src/hooks/useDialog";
 
 export const decorators = [
   (Story) => (
     <BrowserRouter>
       <ThemeContextProvider>
         <NotificationProvider>
-          <Story />
+          <DialogProvider>
+            <Story />
+          </DialogProvider>
         </NotificationProvider>
       </ThemeContextProvider>
     </BrowserRouter>
