+++ app/src/views/DesignSystem/index.ts
@@ -1,1 +1,1 @@
-export * from "./DialogPage";
+export * from "./TableDesign";
